propel_beta/requirements.txt
chromadb==0.5.0
Django==5.0.6
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
pyautogen==0.2.21
PyPDF2==3.0.1
typing_extensions==4.12.0
